adicionado um favicon, nome do produto no titulo da pagina do produto , confirmação caso queira excluir um produto do banco e exclusao de resquisios do antigo banco

+++ script/admin.js
@@ -312,8 +312,12 @@ function voltarProduto() {
 
 // ❌ Deletar produto
 async function deletarProduto(id) {
-  await fetch(`${url}/${id}`, { method: "DELETE" });
-  carregarProdutos();
+  if (!confirm("Tem certeza que deseja excluir este produto?")) {
+    return; // Cancela a exclusão se o usuário não confirmar
+  }else{
+    await fetch(`${url}/${id}`, { method: "DELETE" });
+    carregarProdutos();
+  }
 }
 
 // ➕ Adicionar cor
